+++ src/pybamm/solvers/c_solvers/idaklu/Options.cpp
@@ -1,6 +1,7 @@
 #include "Options.hpp"
 #include <iostream>
 #include <stdexcept>
+#include <cmath>
 
 
 using namespace std::string_literals;
@@ -11,9 +12,25 @@ SetupOptions::SetupOptions(py::dict &py_opts)
       precon_half_bandwidth(py_opts["precon_half_bandwidth"].cast<int>()),
       precon_half_bandwidth_keep(py_opts["precon_half_bandwidth_keep"].cast<int>()),
       num_threads(py_opts["num_threads"].cast<int>()),
+      num_solvers(py_opts["num_solvers"].cast<int>()),
       linear_solver(py_opts["linear_solver"].cast<std::string>()),
       linsol_max_iterations(py_opts["linsol_max_iterations"].cast<int>())
 {
+    if (num_solvers > num_threads)
+    {
+        throw std::domain_error(
+            "Number of solvers must be less than or equal to the number of threads"
+        );
+    }
+
+    // input num_threads is the overall number of threads to use. num_solvers of these
+    // will be used to run solvers in parallel, leaving num_threads  / num_solvers threads
+    // to be used by each solver. From here on num_threads is the number of threads to be used by each solver
+    num_threads = static_cast<int>(
+        std::floor(
+            static_cast<double>(num_threads) / static_cast<double>(num_solvers)
+        )
+    );
 
     using_sparse_matrix = true;
     using_banded_matrix = false;