+++ tests/unit/test_spatial_methods/test_finite_volume/test_extrapolation.py
@@ -9,7 +9,6 @@ from tests import (
     get_1p1d_mesh_for_testing,
 )
 import numpy as np
-import unittest
 
 
 def errors(pts, function, method_options, bcs=None):
@@ -57,7 +56,7 @@ def get_errors(function, method_options, pts, bcs=None):
     return l_errors, r_errors
 
 
-class TestExtrapolation(unittest.TestCase):
+class TestExtrapolation:
     def test_convergence_without_bcs(self):
         # all tests are performed on x in [0, 1]
         linear = {"extrapolation": {"order": "linear"}}
@@ -262,8 +261,8 @@ class TestExtrapolation(unittest.TestCase):
 
         # check constant extrapolates to constant
         constant_y = np.ones_like(macro_submesh.nodes[:, np.newaxis])
-        self.assertEqual(extrap_left_disc.evaluate(None, constant_y), 2)
-        self.assertEqual(extrap_right_disc.evaluate(None, constant_y), 3)
+        assert extrap_left_disc.evaluate(None, constant_y) == 2
+        assert extrap_right_disc.evaluate(None, constant_y) == 3
 
         # check linear variable extrapolates correctly
         linear_y = macro_submesh.nodes
@@ -297,7 +296,7 @@ class TestExtrapolation(unittest.TestCase):
 
         # check constant extrapolates to constant
         constant_y = np.ones_like(micro_submesh.nodes[:, np.newaxis])
-        self.assertEqual(surf_eqn_disc.evaluate(None, constant_y), 1.0)
+        assert surf_eqn_disc.evaluate(None, constant_y) == 1.0
 
         # check linear variable extrapolates correctly
         linear_y = micro_submesh.nodes
@@ -359,7 +358,7 @@ class TestExtrapolation(unittest.TestCase):
         np.testing.assert_array_almost_equal(
             extrap_flux_left_disc.evaluate(None, constant_y), 0
         )
-        self.assertEqual(extrap_flux_right_disc.evaluate(None, constant_y), 0)
+        assert extrap_flux_right_disc.evaluate(None, constant_y) == 0
 
         # check linear variable extrapolates correctly
         np.testing.assert_array_almost_equal(
@@ -448,7 +447,7 @@ class TestExtrapolation(unittest.TestCase):
         extrap_right = pybamm.BoundaryValue(var, "right")
         disc.set_variable_slices([var])
         extrap_right_disc = disc.process_symbol(extrap_right)
-        self.assertEqual(extrap_right_disc.domain, ["negative electrode"])
+        assert extrap_right_disc.domain == ["negative electrode"]
         # evaluate
         y_macro = mesh["negative electrode"].nodes
         y_micro = mesh["negative particle"].nodes
@@ -462,7 +461,7 @@ class TestExtrapolation(unittest.TestCase):
         extrap_right = pybamm.BoundaryValue(var, "right")
         disc.set_variable_slices([var])
         extrap_right_disc = disc.process_symbol(extrap_right)
-        self.assertEqual(extrap_right_disc.domain, [])
+        assert extrap_right_disc.domain == []
 
         # 2d macroscale
         mesh = get_1p1d_mesh_for_testing()
@@ -471,7 +470,7 @@ class TestExtrapolation(unittest.TestCase):
         extrap_right = pybamm.BoundaryValue(var, "right")
         disc.set_variable_slices([var])
         extrap_right_disc = disc.process_symbol(extrap_right)
-        self.assertEqual(extrap_right_disc.domain, [])
+        assert extrap_right_disc.domain == []
 
         # test extrapolate to "negative tab" gives same as "left" and
         # "positive tab" gives same "right" (see get_mesh_for_testing)
@@ -497,13 +496,3 @@ class TestExtrapolation(unittest.TestCase):
             extrap_pos_disc.evaluate(None, constant_y),
             extrap_right_disc.evaluate(None, constant_y),
         )
-
-
-if __name__ == "__main__":
-    print("Add -v for more debug output")
-    import sys
-
-    if "-v" in sys.argv:
-        debug = True
-    pybamm.settings.debug_mode = True
-    unittest.main()